chore: cleanup library for consistency
